api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.5
websockets==12.0
